update from libtiff CVS
git-svn-id: https://svn.osgeo.org/gdal/trunk@9429 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/gtiff/libtiff/tif_close.c
@@ -1,4 +1,4 @@
-/* $Header: /cvs/maptools/cvsroot/libtiff/libtiff/tif_close.c,v 1.8 2004/01/30 20:22:18 dron Exp $ */
+/* $Id: tif_close.c,v 1.10 2006/03/25 03:09:24 joris Exp $ */
 
 /*
  * Copyright (c) 1988-1997 Sam Leffler
@@ -54,27 +54,27 @@ TIFFCleanup(TIFF* tif)
 	TIFFFreeDirectory(tif);
 
 	if (tif->tif_dirlist)
-	    _TIFFfree(tif->tif_dirlist);
-	    
+		_TIFFfree(tif->tif_dirlist);
+
 	/* Clean up client info links */
 	while( tif->tif_clientinfo )
 	{
-	    TIFFClientInfoLink *link = tif->tif_clientinfo;
+		TIFFClientInfoLink *link = tif->tif_clientinfo;
 
-	    tif->tif_clientinfo = link->next;
-	    _TIFFfree( link->name );
-	    _TIFFfree( link );
+		tif->tif_clientinfo = link->next;
+		_TIFFfree( link->name );
+		_TIFFfree( link );
 	}
 
 	if (tif->tif_rawdata && (tif->tif_flags&TIFF_MYBUFFER))
-	    _TIFFfree(tif->tif_rawdata);
+		_TIFFfree(tif->tif_rawdata);
 	if (isMapped(tif))
-	    TIFFUnmapFileContents(tif, tif->tif_base, tif->tif_size);
+		TIFFUnmapFileContents(tif, tif->tif_base, tif->tif_size);
 
 	/* Clean up custom fields */
-	if (tif->tif_nfields > 0) 
+	if (tif->tif_nfields > 0)
 	{
-	    int  i;
+		size_t  i;
 
 	    for (i = 0; i < tif->tif_nfields; i++) 
 	    {
